Corrected pathing (mostly) and the Grid offset.

+++ core/src/com/quickbite/economy/util/TimeOfDay.kt
@@ -18,7 +18,7 @@ object TimeOfDay{
     val timeScale = 5
 
     /** Takes into account the timeScale of delta moveTime from TimeUtil. This is the calculated timeScale of moveTime progression*/
-    val currScaledTime:Int
+    val currScaledTime:Float
         get() = timeScale *TimeUtil.deltaTimeScale
 
     var day = 0
